Split popup and popup manager out in to separate files in global-popup
